+++ resources/mail/passwordResetToken-html.php
@@ -7,7 +7,7 @@ use yii\helpers\Html;
 /** @var \app\models\User $user */
 /** @var \yii\web\View $this */
 
-$resetLink = Yii::$app->urlManager->createAbsoluteUrl(['site/reset-password', 'token' => $user->password_reset_token]);
+$resetLink = Yii::$app->urlManager->createAbsoluteUrl(['user/reset-password', 'token' => $user->password_reset_token]);
 ?>
 <div class="password-reset">
     <p>Hello <?= Html::encode($user->username) ?>,</p>